Refactor random to split out base and compat code
